% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{ark_datasets}
\alias{ark_datasets}
\title{List known Arkansas GIS datasets with friendly names}
\usage{
ark_datasets()
}
\value{
tibble: key, service, layer_id, name
}
\description{
List known Arkansas GIS datasets with friendly names
}
