% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{ark_hospitals}
\alias{ark_hospitals}
\title{Get Arkansas hospital/health facility locations}
\usage{
ark_hospitals(limit = 1000)
}
\arguments{
\item{limit}{Max records (default 1000)}
}
\value{
tibble with hospital attributes
}
\description{
Get Arkansas hospital/health facility locations
}
