% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{ark_layers}
\alias{ark_layers}
\title{List layers in an Arkansas GIS feature service}
\usage{
ark_layers(service)
}
\arguments{
\item{service}{Service key from ark_services() or full path}
}
\value{
tibble: id, name, type
}
\description{
List layers in an Arkansas GIS feature service
}
