% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{ark_senate_districts}
\alias{ark_senate_districts}
\title{Get Arkansas Senate district boundaries}
\usage{
ark_senate_districts(limit = 100)
}
\arguments{
\item{limit}{Max records (default 100)}
}
\value{
tibble with senate district attributes
}
\description{
Get Arkansas Senate district boundaries
}
