% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ark.R
\name{ark_services}
\alias{ark_services}
\title{List available Arkansas GIS feature services}
\usage{
ark_services()
}
\value{
tibble: key, service_path
}
\description{
List available Arkansas GIS feature services
}
